Refactor data type implementations to use their own file each
