+++ src/sqlalchemy_cratedb/type/__init__.py
@@ -1,0 +1,3 @@
+from .array import ObjectArray
+from .geo import Geopoint, Geoshape
+from .object import ObjectType
