+++ tests/compiler_test.py
@@ -37,8 +37,7 @@ try:
 except ImportError:
     from sqlalchemy.ext.declarative import declarative_base
 
-from sqlalchemy_cratedb import SA_VERSION, SA_1_4, SA_2_0
-from sqlalchemy_cratedb import ObjectType
+from sqlalchemy_cratedb import SA_VERSION, SA_1_4, SA_2_0, ObjectType
 from crate.client.test_util import ParametrizedTestCase
